requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirments.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via gemini (pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.115.12
    # via gemini (pyproject.toml)
fastapi-cli==0.0.7
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
jiter==0.9.0
    # via openai
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
openai==1.69.0
    # via gemini (pyproject.toml)
pydantic==2.11.0
    # via
    #   fastapi
    #   openai
pydantic-core==2.33.0
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.0
    # via fastapi-cli
ruff==0.11.2
    # via gemini (pyproject.toml)
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.46.1
    # via fastapi
tqdm==4.67.1
    # via openai
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.0
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
uvicorn==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
